feat: 반품 규정 삭제 기능 구현

+++ src/main/java/store/mybooks/resource/return_rule/repository/ReturnRuleRepository.java
@@ -15,7 +15,7 @@ import store.mybooks.resource.return_rule.entity.ReturnRule;
  * -----------------------------------------------------------<br>
  * 2/21/24        minsu11       최초 생성<br>
  */
-public interface ReturnRuleRepository extends JpaRepository<ReturnRule, Integer>, ReturnRuleRepositoryCustom {
+public interface ReturnRuleRepository extends JpaRepository<ReturnRule, Long>, ReturnRuleRepositoryCustom {
 
     Optional<ReturnRule> findByReturnRuleNameId(String id);
 }
